Fixed duplicate tag bug for eureka selector view controller.
